FSJ & Afrika Korps revision
In updating the Afrika Korps catalog to be more user and maintainer friendly, I've done a complete overhaul from the ground up in accordance with how other systems handle their data.

In this way, with a simple set of guidelines, we can move forward and update the rest of the data if this format is found to be conducive to co-operation and quality assurance.

I also added the Fallschirmjager splatbook "Death from Above" that can be added to any Mid-War German book; right now, it's only implemented in Afrika Korps since I don't see Ghost Panzers or Iron Cross in Battlescribe right now.