Like a Ferdinand at Kursk, this catalog's Finished
Hey there, AncestralHeritage; I'm real sorry for gutting your work like this, but I feel it was for the best in terms of useability.
Adding references to everything is nice, but it clogs up the actual UI and makes every entry a pain to parse.
Naming every ability (Confident 4+) vs just (4+) doesn't add anything to the useability of the catalog, either.
Knowing which side it came from (Enemy at the Gates or Red Banner) doesn't really matter either since they're in the same book now and can be used together to my knowledge.
Adding the model kit number to every unit was nice when it was correct, but Battlefront has changed some names and completely gotten rid of others; long and short, most of the SKUs are now useless.
Finally, making everything an SEG isn't ideal, either; it adds an option that users can click that does actually nothing. If it's MinMax1 and inside an SEG, it's not an option at all.

I apologize if this seems like I'm stomping all over your work, but I really wish you'd looked at how literally any of the other catalogs were already formatted before doing this one; please understand that I appreciate your work and am willing to discuss this with you at length if you so wish.